fix: P4ADEV-3829 CVE fix

+++ src/main/java/it/gov/pagopa/pu/debtpositions/repository/DebtPositionTypeOrgOperatorsRepository.java
@@ -18,12 +18,12 @@ public interface DebtPositionTypeOrgOperatorsRepository extends JpaRepository<De
   List<DebtPositionTypeOrgOperators> findByDebtPositionTypeOrgId(Long debtPositionTypeOrgId);
 
   @Transactional
-  long deleteByDebtPositionTypeOrgId(Long debtPositionTypeOrgId);
+  Integer deleteByDebtPositionTypeOrgId(Long debtPositionTypeOrgId);
   @Transactional
   @Modifying
   @Query("DELETE FROM DebtPositionTypeOrgOperators dptoo WHERE dptoo.debtPositionTypeOrgId = :debtPositionTypeOrgId "
     + " AND dptoo.operatorExternalUserId IN :operatorExternalUserIds ")
-  int deleteByDebtPositionTypeOrgIdAndOperatorExternalUserId(Long debtPositionTypeOrgId, Set<String> operatorExternalUserIds);
+  Integer deleteByDebtPositionTypeOrgIdAndOperatorExternalUserId(Long debtPositionTypeOrgId, Set<String> operatorExternalUserIds);
 
   List<DebtPositionTypeOrgOperators> findByOperatorExternalUserId(String operatorExternalUserId);
 }
